Sprites are mostly working
I've overhauled the Sprite class to actually properly pull an image
during level loads and then add an image to the sprite during sprite
creation.

What doesn't work at this time is positioning.  They're all defaulting
to 0,0 on the grid which is obviously non-optimal. The code for
displaying them is also not in the game.py class.  This code should be
modified per the changes in this commit from level.py's main function.

Once I have working sprites and a working player character, I can start
adding command input to the main game loop and moving the player
character around.