Player Class working, level display working
Got the player class to initialize and  to display, finally.  The code
to handle both sprites and player are currently only in the level.py
main function as that was an easy place to test the base logic and make
sure level displays work.

Next task is to update the game class to handle the sprites, then I can
proceed on adding input handling to the current gaming loop allowing for
movement and the beginnings of an actual game.