Added collision with walls and blocking tiles
To make the maps actually make sense I added the logic to check if a
tile you're about to move onto is a wall or other blocking type of tile
based off the map configuration.  This should make the movement a bit
more realistic and if walled properly keep the player from moving off
the defined screen.

Also updated the readme to show the current state and update the todo
list.