Organizational Refactor
I figured the best way to get refamiliarized with the project was to
refactor it to be in a slightly better state.  I started by organizing it
into functional modules.

I would like to look at other refactors to make it more effective in coming
commits.  Whenever those may happen.